first attempt of Model_import, but in heavy use of a refactor.
Currently the IR type system that we have for tensors, is just pain to
work with, we should definitely be using something more ergnomic
a tensor should be a type which fully abstracts away from the underlying
type, otherwise we end up with a elseif chain hell for all operations,
no fun.

Also it's probably a bad idea, to handle attributes and values the same
way, attributes can be parsed competely ahead of time, and for values
from intializers we are probably still interessted in the raw data, but
maybe there is a way to model a container which can provide both
interfaces than it would be nice to reuse it.

Also lists are not tensors, that should be explicit!

Probably a good idea to first decide what are my top level types and
reduce that amount to the absolute minimum then we can move onto
implementing all ops in a hopefully decently lean way, the current stuff
is a complete ChatGPT, mess but it was still useful to recognize the
shortcommings of this design.