Big refactor of import Tensor types.
They now fully abstract away from the underlying type including Sym.
Also remodells Shapes and introduces Views with strides for HostTensors.

Makes a big distinction between HostTensors and DeviceTensors, where
operation can immediately be performed on HostTensors (in most cases)
and DeviceTensor operation will have to be done during the runtime.

HostTensors are modelled similar to numpy tensors, with strides and so
on.

There are definitely some big bugs in the current operation
implementations. That will HAVE TO BE fixed, but progress is good