Early LinkedGraph implementation.
Our current hypergraph suite does not provide a easy interface
for graph editing like we would need for something like fusion passes
or construction of a supergraph.

LinkedGraph is a implementation of a hypergraph, where
nodes and edges are seperately allocated, and reference counted.
It only works with DAGs, no need to check for cycles they should never
really happen in our computational graphs.

LinkedGraph is aggressivly optimized for graph editing, so
although concentually each node and edge can be thought of as
ref counted and allocated, we really avoid heap allocations in
most cases and instead pool most allocations.

addNode: O(1)
addEdge: O(|srcs|), basically O(1) because |srcs| is almost always less than 3.
removeNode: O(|N|) worst case, but just because we potentially have to
deallocate all nodes, if we remove the last referenced node.
In case we replace an edge section like
A -> B -> C with A -> C, this results in the minal overhead,
which is a pooled deallocation of B and the edges and a pooled
allocation of the new edge.
removeEdge: O(|N|) worst case, but again same argument as for removeNode.

The implementation is probably big time overkill, because all of the
runtimes are not amortized, so no vectors, we actually use doubly linked
list for stuff like incoming and outgoing edge lists, which may seem
weird, but the idea is that because we pool allocations anyway we still
get decent cache performance and iteration is far from the most
import operation for editing.

For the final implementation pass, where we build the supergraph,
we will most likely use a combination of a ConstGraph for blazily fast
iteration (over the ops applied to a tensor)
and a LinkedGraph for construction, maybe a AdjGraph also works we will
see.