Build API of LinkedGraph.
The big picture for a LinkedGraph becomes a little bit cleaner now.
It a hypergraph representation, which is aggresivly optimized for
editing, over iterators. This alone would not be complex, what makes it
complex that the LinkedGraph removes unreachable nodes recursivly.

So for example if we have a graph:
A -> B
A -> C
B -> C

Let's say we hold a reference somewhere to the Node A.
And we remove the edge: A -> B, from the graph.
This will make node B unreachable, which means it get's deallocated as
well. This also then implies that the edge B -> C is invalid so we
deallocate it as well.

This is especially difficult because a LinkedGraph allows forwards and
backwards iteration, where we additionally hold backlinks
(i.e. we can traverse a edge A -> B from B to A).

We can do this while iterating over outgoing / incoming edges of
any node, while not losing the iterator state.
This is perfect for traversing a graph with DFS,
while editing (i.e. replacing) segments.

The same invariant is true for inserting new edges into the graph,
this is a luxery that we would never be able to reach with a
standard implementation with something like a std::vector +
std::shared_ptr.

Our implementation aggressivly uses doublely linked rings for
incoming / outgoing edges of nodes, which makes iteration stable
with some exceptions.
Similarly the srcs nodes of a edge are stored in a linked list, but
here a forward list is actually enough.

Linked list generally are a horrible idea, but the hole idea here is
that all linked list nodes (outgoing, incoming, srcs) have the same byte
size and are allocated from the same freelist pool. Additionally
we store pointers to list entries, which allows constant time removal
from outgoing and incoming (ignoring the memory managment overhead).

Now i feel like i know a bit more what this is useful for:
- Iam not 100% sure if this is a good datastructure for building the
  supergraph, where something like a AdjGraph + ConstGraph might
  actually be better, where ConstGraph holds the underlying operations
  between tensors and AdjGraph holds the implementation edges,
  while both shared the same nodes (NodeId).
- It might still be a choice for a building the supergraph, because if
  we want to abstract over our shader implementations which we
  definitely want we have to find a decent api for viewing the graph
  structure, at this point the by far most friendly interface is the
  LinkedGraph, but ofcause it comes with it's performance hits.
- What the LinkedGraph is definitely really good for is for
  the trivial fusion pass. It's quite easy to define a simple
  abstraction for a fusion rule here, and because fusion generally only
  replaces edges (which is more expensive in a AdjGraph) and
  never increases the overall edge count, we can perform the complete
  editing without calling malloc, this is a beautiful properly,
  given the complexity and the dynamic nature of a linked graph.

Maybe at some point i will benchmark those graphs, would be really
interessting if a linked structure can actually beat a
vector based structure like a AdjGraph in certain applications,
like editing or if linked list are actually just soo
slow that it's really really never worth it even if you try to
super hard to make them work.