General cano + spec + dce + freeze structure
Dead Code Elimination (DCE)

We use the memory managmenet property of the LinkedGraph here.
First we make sure that we only hold a reference to the input node,
with this we are guarnateed that any node, which is not reachable
from the input node is deallocated implicity.
In dce we now construct a adj graph from a linked graph by traversing
the graph backwards from the output node and only taking incoming edges.
This makes sure that we only consider points, which can contribute to
the output, all other nodes are dropped implicitly.

Hypergraph pattern matching!!

Even more fun interessting subprojects =^).
Jokes aside we really really needed this, we now have a pretty
nice way to match pattern within a graph.

We first have to define a GraphPattern,
which you can imagine itself to be some form of graph structure,
but ofcause a lot more simple (i.e. just a shared_ptr maskacer)

Then we match this pattern again the graph.
The function match_first is quite simple, but it will return the same
match multiple times.

Ofcause we are interessted in all matches in most cases, so we also
added match_all which returns a "rangeish" thing, which we can iterate
over to get all matches.
What we actually get is a generator, which means that we lazily evaluate
the matches and resume the internal matching engine at the point where
we last stopped.
Now for a ConstGraph without any mutations this is a bit overkill, but
was interessting to started with coroutines with this.

The hole idea behind using couroutines here, is that we also wanted to
support mutations of the graph, while matching the graph.
We support this type of pattern only for a LinkedGraph, which has
decently stable iterators. With the current implementation for
LinkedGraph we can now iterate over all matches and modifiying the
graph while handling matches. Any modifications done here will also be
considered within the same coroutines. This is quite a nice
pattern to collapse patterns like
A -> B -> C -> D directly into A -> D, while only specifying the
rule X -> Y -> Z = X -> Z.

Now we finally have proper datastructures to implement the
cano pass.

For the later superpass building pass we will most likely use the
ConstGraph path because from the operations graphs perspective no
mutations actually happend and matching on ConstGraphs is significantly
more performant.