Dead Code Elimination (DCE)
We use the memory managmenet property of the LinkedGraph here.
First we make sure that we only hold a reference to the input node,
with this we are guarnateed that any node, which is not reachable
from the input node is deallocated implicity.
In dce we now construct a adj graph from a linked graph by traversing
the graph backwards from the output node and only taking incoming edges.
This makes sure that we only consider points, which can contribute to
the output, all other nodes are dropped implicitly.