yay first trivial fusion rule.
I think the pattern matching API turned out great,
see SliceSlice.hpp and look how beautifully simple it actually looks.

Currently this fusion basically never kicks in because doens't work for
Symbolic values, which they are almost always going to be, we need
min and max symbolic expressions for that.