This worked better than expected xD.
Okay so the interface for Shaders actually looks quite good,
we can now match simple shaders like a simple Conv
or a Conv + Relu on the computational graph and generate the
implementation edges for them.
We implemented this for the BasicPool and BasicUpsample shaders as well.

A current limitation is that we are not able to match upwards, which
means that we cannot properly match operations which have more than one
input like concat.
For example:
(A,B) -> C
We can match the edge, but we actually require the NodePatterns for A
and B to retrive them later when we create the hyperedges in the
implementation graph.

Now we would solve this within the impl.cpp by also allowing edges, but
this is not clean enough because we might end up wanting to match
patterns like:
B -> X
(A,X) -> C
Where B and A are inputs and C is the output, this means that we have to
improve our match_all implementation for ConstGraphs.
I think that for linked graph matching we should never encounter this pattern so
can just throw unimplemented or unreachable here.

Next up: Improving the match_all for ConstGraph