begin with mini symbolic expression compiler
One minor goal here is to minimize the expressions a bit further,
but the more important goal is to find a way to represent
the symbolic expressions as a set of binary operations.

The native representation may include stuff like mul(x,y,z), which is
obviously not binary. The complex part here is to determine if we
should compute (x*y)*z or (y*z)*x or (z*x)*y.
We will later try to select the best option, while ensuring that the
resulting set of operations is "minimal"
(obviously not perfectly minimal, but close enough).