Symbolic expression compiler.
Takes a SymGraph and a set of symbols that we need during
runtime, and generates a SymIR, which contains only binary operations
and removes all symbolic values that are not required during runtime.

This actually helped much more than i thought, almost cuts the amount of
operations required in half i guess this happens because the symbolic
engine sometimes generates symbols just for prove statements and we do
not need those proof symbols during runtime.

At this point this is super simple basically just does a dce pass.
It would be interessting to also search for more optimal paths by
looking at the affine expressions of the symbols.
Because those encode the proofen reductions of the symbolic engines
in a potentially more optimal way.