Add serialization into .dnx file format.
We decided to use flatbuffers, for the dnx file format,
they are pretty easy to parse,
and more importantly have no runtime dependencies.

I also throught about using protobuf for this, but
what annoyed me is that protobuf requires libprotobuf.so
to be loaded at runtime, which or to statically link
libprotobuf.a which isn't a super small library at least
compared to libflatbuffers, where the library is basically just a couple
of headers and no source files.