adding hash to DbOp!
What was missing is something that allows us to seperate between two
different applications of a shader even if they have the same
configuration and matched the same pattern.
For example our DirectConv has the same config regardless the input
and output layouts. We now just compute a big hash over the Tensor
Descriptors hopefully collisions are really really rare.